+++ packages/utils/package.json
@@ -38,6 +38,7 @@
         "lodash.has": "^4.5.2",
         "lodash.set": "^4.3.2",
         "lodash.unset": "^4.5.2",
+        "p-map": "^4.0.0",
         "shallow-clone": "^3.0.1"
     },
     "devDependencies": {
