+++ packages/utils/test/promises-spec.ts
@@ -1,7 +1,7 @@
 import 'jest-extended';
 import { TSError } from '../src/errors';
 import {
-    waterfall, pWhile, pRetry, PRetryConfig, getBackoffDelay, PWhileOptions, pMap, pDelay
+    waterfall, pWhile, pRetry, PRetryConfig, getBackoffDelay, PWhileOptions
 } from '../src/promises';
 
 describe('Utils', () => {
@@ -187,37 +187,5 @@ describe('Utils', () => {
         });
 
         it('should run until it times out', async () => expect(pWhile(async () => false, defaultOptions)).rejects.toThrow(/Request timeout after \d+(ms|sec) while waiting/));
-    });
-
-    describe('pMap', () => {
-        it('should be able to work without concurrency', () => {
-            const items: { int: number }[] = [{ int: 1 }, { int: 3 }];
-            return expect(pMap(items, (item) => {
-                item.int++;
-                return item;
-            })).resolves.toEqual([
-                { int: 2 }, { int: 4 }
-            ]);
-        });
-
-        it('should be able to work with a concurrency of 2', async () => {
-            const items: number[] = [1, 2, 3, 4, 5];
-            const start = Date.now();
-            const concurrency = 2;
-            const wait = 500;
-
-            const result = pMap(items, async (item, i) => {
-                const diff = Date.now() - start;
-                if (i % concurrency === 0) {
-                    await pDelay(wait);
-                }
-                return Math.floor(diff / wait) * wait;
-            }, {
-                concurrency
-            });
-            return expect(result).resolves.toEqual([
-                0, 0, 500, 500, 1000
-            ]);
-        });
     });
 });
